Verify propagation at all recursors
This is necessary for enterprise PKI's where the PKI only talks to a set of DNS-server rather than talking to the authoritative directly.
In this scenario we need Lego to verify the propagation of the record on a user provided set of DNS-servers, or else Lego sees the record only on authoritative servers and OK's the request while the PKI will fail as it's DNS part does not have the record yet.